Fix mkl interface and others.
Fix in the way lapacke is linked.  MKL does not have a lapacke library
like netlib ot others, it has only a header.

Fix also in the _MKL_DSS_HANDLE_t name... not sure when it changed. And
other things related with the complains of intel_oneap_mkl.

Fix malloc(a,b) entrees, replacing with malloc(a*b)